+++ guillotina/exceptions.py
@@ -53,8 +53,12 @@ class PreconditionFailed(Exception):
 
 
 class RequestNotFound(Exception):
-    """Lookup for the current request for request aware transactions failed
+    """Lookup for the current request failed
     """
+
+
+class TaskNotFound(Exception):
+    pass
 
 
 @implementer(IUnauthorized)